+++ src/NeuralNet/Layers/Multiclass.php
@@ -181,7 +181,9 @@ class Multiclass implements Output
         $fanOut = count($this->classes);
 
         $w = $this->weightInitializer->initialize($fanIn, $fanOut);
-        $b = $this->biasInitializer->initialize(1, $fanOut)->columnAsVector(0);
+
+        $b = $this->biasInitializer->initialize(1, $fanOut)
+            ->columnAsVector(0);
 
         $this->weights = new MatrixParam($w);
         $this->biases = new VectorParam($b);