+++ tests/Datasets/Generators/HyperplaneTest.php
@@ -3,23 +3,23 @@
 namespace Rubix\ML\Tests\Datasets\Generators;
 
 use Rubix\ML\Datasets\Dataset;
-use Rubix\ML\Datasets\Unlabeled;
-use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\Datasets\Labeled;
+use Rubix\ML\Datasets\Generators\Hyperplane;
 use Rubix\ML\Datasets\Generators\Generator;
 use PHPUnit\Framework\TestCase;
 
-class BlobTest extends TestCase
+class HyperplaneTest extends TestCase
 {
     protected $generator;
 
     public function setUp()
     {
-        $this->generator = new Blob([0, 0], 1.0);
+        $this->generator = new Hyperplane([0, 0], 0.0);
     }
 
     public function test_build_generator()
     {
-        $this->assertInstanceOf(Blob::class, $this->generator);
+        $this->assertInstanceOf(Hyperplane::class, $this->generator);
         $this->assertInstanceOf(Generator::class, $this->generator);
 
         $this->assertEquals(2, $this->generator->dimensions());
@@ -29,7 +29,7 @@ class BlobTest extends TestCase
     {
         $dataset = $this->generator->generate(30);
 
-        $this->assertInstanceOf(Unlabeled::class, $dataset);
+        $this->assertInstanceOf(Labeled::class, $dataset);
         $this->assertInstanceOf(Dataset::class, $dataset);
 
         $this->assertCount(30, $dataset);